refactor: 拆分eisland多线程下载引擎
