+++ src/main/core/downloadEngine/config/index.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_THREADS = 8;
+export const MAX_THREADS = 16;
+export const MIN_THREADS = 1;
+export const EMIT_INTERVAL_MS = 160;
+export const MIN_CHUNK_BYTES = 1024 * 1024;
